+++ src/celebrimbor/__init__.py
@@ -34,7 +34,7 @@ if TYPE_CHECKING:
 
     from .context import Context
 
-__version__ = "0.7.1"
+__version__ = "0.8.0"
 
 __all__ = [
     "CheckResult",